Added option to reduce visual scale of diagrams
mermaid_img_scale_reduction uses CSS height and width
to scale down the visual size of rendered mermaid diagrams.
This is a direct counter to puppeteer scale fed to
mermaid-cli so that setting both to the same
integer larger than 1 will result in a higher resolution
diagram at the natural 'scale 1' visual size.