fix AttributeError caused by unset self._cover_title
When the style_for_print function is called, the cover_title property of the Options class is accessed, which in turn accesses the _cover_title attribute of the instance. This attribute was only available when cover is True, but the style_for_print function is also called when cover is False.