fix: avoid `AttributeError` caused by unset `self._cover_title` when `cover` is False (#83)
* fix AttributeError caused by unset self._cover_title

When the style_for_print function is called, the cover_title property of the Options class is accessed, which in turn accesses the _cover_title attribute of the instance. This attribute was only available when cover is True, but the style_for_print function is also called when cover is False.

* add defaults values to conditional instantiations

Values for _cover_subtitle and _cover_logo were only assigned conditionally before. This could end in AttributeErrors, therefore added a default value.